feat: 라이어 게임 방 관리 시스템 구현
### API 서버
- 방 생성 (POST /api/rooms) - 인증 필요, 역할 기반 접근 제어
- 방 목록 조회 (GET /api/rooms) - 인증 불필요, 상태 필터링
- 방 코드 조회 (GET /api/rooms/:code) - 인증 불필요
- Room Entity 및 관련 DTO, Service, Controller 구현
- Supabase JWT 인증 전략 통합
- User/Room Entity ID 타입 통합 (unsigned int)

### 웹 프론트엔드
- 실제 API 연동을 통한 방 목록 조회
- 방 생성 기능 (인증 필요)
- 방 코드로 참가 모달 구현
- useRooms 커스텀 훅으로 상태 관리
- API 클라이언트 및 타입 정의
- 반응형 UI 개선

### 기술 개선
- TypeScript 타입 안정성 강화
- 에러 핸들링 및 로딩 상태 관리
- 개발 환경 API URL 설정 (.env.local)

+++ apps/api/src/user/entities/user.entity.ts
@@ -23,7 +23,7 @@ export enum UserRole {
 @Index('idx_users_email', ['email'])
 @Index('idx_users_tier', ['tier'])
 export class UserEntity {
-  @PrimaryGeneratedColumn('increment', { type: 'bigint', unsigned: true })
+  @PrimaryGeneratedColumn('increment', { type: 'int', unsigned: true })
   id: number;
 
   @Column({ type: 'varchar', length: 255 })
